# backend/swagger.yaml
openapi: 3.0.0
info:
  title: QR Ordering System API
  description: |
    Complete API for a QR-based restaurant ordering system with authentication, product management, and order processing.

    ## Features
    - **Authentication**: User registration, login, and JWT-based authentication
    - **Products**: Browse products by category (Burgers, Hotdogs, Drinks)
    - **Orders**: Create and manage orders with dine-in or takeout options
    - **Cart**: Client-side cart management with order submission

    ## Authentication
    Some endpoints require authentication using Bearer tokens. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@qrordering.com
servers:
  - url: http://localhost:5000
    description: Local development server
paths:
...
  /products/{id}:
    get:
      summary: Get product by ID
      description: Retrieve a single product by its ID. No authentication required.
      operationId: getProductById
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: Product ID
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product not found"
  /products/category/{category}:
    get:
      summary: Get products by category
      description: Retrieve all products in a specific category. No authentication required.
      operationId: getProductsByCategory
      tags:
        - Products
      parameters:
        - in: path
          name: category
          required: true
          description: Product category
          schema:
            type: string
            enum: [Burgers, Hotdogs, Drinks]
          example: "Burgers"
      responses:
        "200":
          description: An array of products in the specified category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  # Orders Endpoints
  /orders:
    get:
      summary: Get all orders
      description: Retrieve a list of all orders with their items.
      operationId: getAllOrders
      tags:
        - Orders
      responses:
        "200":
          description: An array of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"

    post:
      summary: Create a new order
      description: Create a new order with order items. No authentication required.
      operationId: createOrder
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Invalid order data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid order data"
  /orders/{id}:
    get:
      summary: Get order by ID
      description: Retrieve a single order with its items by ID.
      operationId: getOrderById
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          description: Order ID
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Order details with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Order not found"

    put:
      summary: Update an order
      description: Update order status and other details.
      operationId: updateOrder
      tags:
        - Orders
      parameters:
        - in: path
          name: id
          required: true
          description: Order ID
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Order not found"

  # Accounts Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Classic American Burger"
        description:
          type: string
          example: "A classic American-style burger with all the traditional fixings"
        price:
          type: number
          format: decimal
          example: 10.99
        category:
          type: string
          enum: [Burgers, Hotdogs, Drinks]
          example: "Burgers"
        image_url:
          type: string
          example: "/images/classic-american-burger.jpg"
        is_available:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"

    # Order Schemas
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        order_id:
          type: integer
          example: 1
        product_id:
          type: integer
          example: 1
        product_name:
          type: string
          example: "Classic American Burger"
        quantity:
          type: integer
          example: 2
        unit_price:
          type: number
          format: decimal
          example: 10.99
        subtotal:
          type: number
          format: decimal
          example: 21.98
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        order_number:
          type: string
          example: "ORD-1705318200000-123"
        table_number:
          type: string
          nullable: true
          example: "5"
        customer_name:
          type: string
          nullable: true
          example: "John Doe"
        total_amount:
          type: number
          format: decimal
          example: 45.97
        order_type:
          type: string
          enum: [dine_in, takeout]
          example: "dine_in"
        order_status:
          type: string
          enum: [pending, completed, cancelled]
          example: "pending"
        payment_method:
          type: string
          enum: [debit_card, credit_card, pay_at_counter]
          example: "pay_at_counter"
        payment_status:
          type: string
          enum: [pending, paid, failed]
          example: "pending"
        payment_notes:
          type: string
          nullable: true
          example: null
        notes:
          type: string
          nullable: true
          example: "No onions please"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    OrderWithItems:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            orderItems:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
    CreateOrderRequest:
      type: object
      required:
        - order_type
        - total_amount
        - payment_method
        - items
      properties:
        table_number:
          type: string
          description: Required for dine_in orders
          example: "5"
        customer_name:
          type: string
          description: Required for takeout orders
          example: "John Doe"
        total_amount:
          type: number
          format: decimal
          example: 45.97
        order_type:
          type: string
          enum: [dine_in, takeout]
          example: "dine_in"
        payment_method:
          type: string
          enum: [debit_card, credit_card, pay_at_counter]
          example: "pay_at_counter"
        notes:
          type: string
          nullable: true
          example: "No onions please"
        items:
          type: array
          items:
            type: object
            required:
              - product_id
              - product_name
              - quantity
              - unit_price
              - subtotal
            properties:
              product_id:
                type: integer
                example: 1
              product_name:
                type: string
                example: "Classic American Burger"
              quantity:
                type: integer
                example: 2
              unit_price:
                type: number
                format: decimal
                example: 10.99
              subtotal:
                type: number
                format: decimal
                example: 21.98
    UpdateOrderRequest:
      type: object
      properties:
        order_status:
          type: string
          enum: [pending, completed, cancelled]
          example: "completed"
        payment_status:
          type: string
          enum: [pending, paid, failed]
          example: "paid"
        payment_notes:
          type: string
          nullable: true
          example: "Payment completed successfully"
        notes:
          type: string
          nullable: true
          example: "Order ready for pickup"
...